Report
HRM WEBSIDE TEST 
– AUTOMATIC TEST BY SELENIUM
Group Member: Sachin, Shiyu Fu
Abstract/Overview:
We use selenium to test an HRM website, including automatic use admin user name to sign in the website, verified all menu buttons exist and all in the correct position on home page, automatic add a new employee to website member, finally, take a screen shot to all user page, and extract all user information into an excel document, then, lot out.
Reflects:
At the first step, we choose this topic that is using selenium to test an HRM website, and we learn what the selenium is, how we can install all the selenium followed steps, and then we downloaded all components and installed them to our laptop. After that, we wrote a plan including: how many days per week we work together in library, how many tasks we need to complete for this project, and how many times we need to complete for each task. We spend one month to complete this step. At this step, we went very well.
At the second step, we spend two weeks to learn how to use selenium to write code, then we start to code login function and log out function. We define the button position and define the admin username and password to automatic sign in and sign out. We just spend one week to complete it. We went well for this part. Then we try to write a function that is used for verified the menu button position in home page all are correct, because for our app, if we want to do or want to check anything, firstly we need to ensure everything and every button is ok in this website. We spend one week to done that, we went very well for this part. After that we got our first challenge that is we can not create new user, we spend one week but we didn’t solve it because we discover user in this website is full, so we change our idea to create a new employee, but we got another challenge that is adding a employee photo to employee information. We spend two week to find a way that is using AutoIt app to add the photo, so we complete it so hard. Finally, we write a code to take a screen shot and create a new Excal and copy all employee information to Excal. We spend two weeks to complete it, it’s a small challenge for us. Then, we finish our second step.
At the third step, we spend two weeks to check and test our code if everything is ok or not, we corrected some bugs in our code, we summarized all of our targets are complete, and also we summarized our difficult problems in our project, and we summarized the process for our whole project, then we finish our project.